Update bootstrap and manifest for v23->jiri tool rename
This updates the bootstrap script and the minimal manifest to build the
'jiri' tool which is the new name of the repository management tool.
This also updates the 'default' manifest to import minimal and contain
the dependencies of the tool (previously the default manifest did not
include x.lib, which jiri needs).

Change-Id: I6c18e88c9f6e94155930147e008b2764eb64c2fe